requirements.txt
deepeval
langchain
langchain-community
langchain-ollama
pandas
python-dotenv
neo4j 
py2neo 
scikit-learn
streamlit
fuzzywuzzy 
python-Levenshtein
matplotlib
seaborn
replicate